修改基础配置，增加产品相关接口

+++ src/main/java/com/atxiaoming/vo/RespBean.java
@@ -10,23 +10,23 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class RespBean {
 
-    private  long code;
-    private  String msg;
+    private Integer code;
+    private String msg;
     private Object data;
 
     public  static RespBean success(){
-        return new RespBean(RespBeanEnum.SUCCESS.getCode(),RespBeanEnum.SUCCESS.getMsg(),null);
+        return new RespBean(RespBeanEnum.SUCCESS.getCode(), RespBeanEnum.SUCCESS.getMsg(),null);
     }
 
     public  static RespBean success(Object obj){
-        return new RespBean(RespBeanEnum.SUCCESS.getCode(),RespBeanEnum.SUCCESS.getMsg(),obj);
+        return new RespBean(RespBeanEnum.SUCCESS.getCode(), RespBeanEnum.SUCCESS.getMsg(),obj);
     }
 
     public  static RespBean error(RespBeanEnum respBeanEnum){
         return new RespBean(respBeanEnum.getCode(),respBeanEnum.getMsg(),null);
     }
 
-    public  static RespBean error(RespBeanEnum respBeanEnum,Object obj){
+    public  static RespBean error(RespBeanEnum respBeanEnum, Object obj){
         return new RespBean(respBeanEnum.getCode(),respBeanEnum.getMsg(),obj);
     }
 }
